Extract withdrawalEvent into the domain package. Building structure here.
