+++ bank-withdrawal-api/src/main/java/com/bank/bankwithdrawalapi/controller/BankAccountController.java
@@ -60,31 +60,3 @@ public class BankAccountController {
     }
 }
 
-public class WithdrawalEvent {
-    private BigDecimal amount;
-    private Long accountId;
-    private String status;
-
-    public WithdrawalEvent(BigDecimal amount, Long accountId, String status) {
-        this.amount = amount;
-        this.accountId = accountId;
-        this.status = status;
-    }
-
-    public BigDecimal getAmount() {
-        return amount;
-    }
-
-    public Long getAccountId() {
-        return accountId;
-    }
-
-    public String getStatus() {
-        return status;
-    }
-
-    // Convert to JSON String
-    public String toJson() {
-        return String.format("{\"amount\":\"%s\",\"accountId\":%d,\"status\":\"%s\"}", amount, accountId, status);
-    }
-}
