+++ bank-withdrawal-api/src/main/java/com/bank/bankwithdrawalapi/domain/WithdrawalEvent.java
@@ -1,0 +1,32 @@
+package com.bank.bankwithdrawalapi.domain;
+
+import java.math.BigDecimal;
+
+public class WithdrawalEvent {
+    private BigDecimal amount;
+    private Long accountId;
+    private String status;
+
+    public WithdrawalEvent(BigDecimal amount, Long accountId, String status) {
+        this.amount = amount;
+        this.accountId = accountId;
+        this.status = status;
+    }
+
+    public BigDecimal getAmount() {
+        return amount;
+    }
+
+    public Long getAccountId() {
+        return accountId;
+    }
+
+    public String getStatus() {
+        return status;
+    }
+
+    // Convert to JSON String
+    public String toJson() {
+        return String.format("{\"amount\":\"%s\",\"accountId\":%d,\"status\":\"%s\"}", amount, accountId, status);
+    }
+}
